Not sure why this wasn't committed

+++ engine/src/AnimationSystem2d.cpp
@@ -27,23 +27,16 @@ namespace core {
 	//TODO these have to do the right thing
 	bool AnimationSystem2d::resetImpl() {
 
-		auto dc = DrawableChange{};
-		dc.operation = DrawableChange::Operation::DESTROY_DRAWABLE;
-		dc.layerId = _drawableLayerId;
-		for (auto& facet : _animationFacets) {
-			dc.facetId = facet.id();
-			single<Renderer>().applyDrawableChange(dc);
-		}
-
-		_movingAnimations.clear();
-		_animationFacets.clear();
-		_animationFacets.shrink_to_fit();
+		if (_facets.reset() != InitStatus::CREATE_TRUE) return false;
 
 		return RenderableSystem2d::resetImpl();
 
 	}
 
 	bool AnimationSystem2d::destroyImpl() {
+
+		if (_facets.destroy() != InitStatus::CREATE_FALSE) return false;
+
 		return RenderableSystem2d::destroyImpl();
 	}
 
@@ -54,42 +47,33 @@ namespace core {
 	bool AnimationSystem2d::handleEvent(AnimationChangeEvent& animationChange) {
 
 	
+		auto facet = _facets.getFacet(animationChange.facetId);
+		if (facet == nullptr) return true;
 
-		for (auto& animation : _animationFacets) {
+			if (facet->currentAnimation.name().compare(animationChange.nextAnimation))  {
 
-			if (animation.of() == animationChange.entity) {
-
-				if (animation.currentAnimation.name().compare(animationChange.nextAnimation))  {
-
-					if (animationChange.endImmediate) {
-						animation.currentAnimation = animation.animationSet.getAnimation(animationChange.nextAnimation);
-						animation.startedAnimation = true;
-					}
-					else {
-						animation.nextAnimation = animationChange.nextAnimation;
-						animation.endAnimation = true;
-					}
-					
+				if (animationChange.endImmediate) {
+					facet->currentAnimation = facet->animationSet.getAnimation(animationChange.nextAnimation);
+					facet->startedAnimation = true;
 				}
-
-				return true;
-
+				else {
+					facet->nextAnimation = animationChange.nextAnimation;
+					facet->endAnimation = true;
+				}
+					
 			}
 
-		}
-		return true;
+			return false;
 
 	}
 
-	AnimationFacet& AnimationSystem2d::createAnimationFacet(Entity& e) {
+	AnimationFacet* AnimationSystem2d::createAnimationFacet(Entity& e) {
 
 		auto newFacet = AnimationFacet{};
 
 		newFacet.setOf(e);
 
-		_animationFacets.push_back(newFacet);
-
-		return _animationFacets.back();
+		return _facets.addFacet(newFacet);		
 
 	}
 
@@ -116,52 +100,28 @@ namespace core {
 
 	std::vector<Facet*> AnimationSystem2d::getFacets(Entity& e) {
 
-		auto out = std::vector<Facet*>{};
-
-		for (auto& facet : _animationFacets) {
-			out.push_back(&facet);
-		}
-
-		return out;
+		return _facets.getBaseFacets(e);
 
 	}
 
 	void AnimationSystem2d::updateImpl(float dt, RuntimeContext& context) {
+		_facets.forEach([&](AnimationFacet* facet)->void {
+			if (facet->isPaused()) return;
+			auto didChangeFrames = facet->currentAnimation.update(context.dt);
+			if (didChangeFrames || facet->startedAnimation) {
 
-		for (auto& facet : _animationFacets) {
-			if (facet.isPaused()) continue;
-			auto didChangeFrames = facet.currentAnimation.update(context.dt);
-			if (didChangeFrames || facet.startedAnimation) {				
-
-				facet.drawable.sourceRect = facet.currentAnimation.frames[facet.currentAnimation.currentFrameIndex].frame;
+				facet->drawable.sourceRect = facet->currentAnimation.frames[facet->currentAnimation.currentFrameIndex].frame;
 				//single<Renderer>().updateDrawable(facet.drawable);
-				facet.startedAnimation = false;
+				facet->startedAnimation = false;
 			}
-		}
+		});
 	}
 
 
 
 	void AnimationSystem2d::destroyFacets(Entity& e) {
 
-		auto movingIt = _movingAnimations.find(e);
-		if (movingIt != std::end(_movingAnimations)) {
-			_movingAnimations.erase(movingIt);
-		}
-
-		auto dc = DrawableChange{};
-		dc.operation = DrawableChange::Operation::DESTROY_DRAWABLE;
-		dc.layerId = _drawableLayerId;
-
-		for (auto it = std::begin(_animationFacets); it != std::end(_animationFacets); ++it) {
-			if (it->of() == e) {
-				
-				dc.facetId = it->id();
-				single<Renderer>().applyDrawableChange(dc);				
-				it = _animationFacets.erase(it);
-				break;
-			}
-		}
+		_facets.removeFacets(e);
 		
 	}
 
@@ -190,29 +150,29 @@ namespace core {
 
 			std::string animationSetName = lua["animationSetName"];
 			
-			auto& facet = system->createAnimationFacet(entity);
+			auto facet = system->createAnimationFacet(entity);
 			auto animationSet = single<TextureManager>().getAnimationSet(animationSetName);
 
-			facet.animationSet = animationSet;
+			facet->animationSet = animationSet;
 
-			facet.currentAnimation = animationSet.getAnimation(currentAnimation);
+			facet->currentAnimation = animationSet.getAnimation(currentAnimation);
 
-			facet.offset = offset.getPixel();
+			facet->offset = offset.getPixel();
 
-			facet.drawable.camera = system->camera();
+			facet->drawable.camera = system->camera();
 
-			facet.drawable.texture = facet.currentAnimation.texture;
+			facet->drawable.texture = facet->currentAnimation.texture;
 
-			facet.drawable.sourceRect = facet.currentAnimation.frames[facet.currentAnimation.currentFrameIndex].frame;
-			facet.drawable.layerId = system->getDrawableLayerId();
-			facet.drawable.targetRect.w = roundFloat(facet.drawable.sourceRect.w * facet.scale.x);
-			facet.drawable.targetRect.h = roundFloat(facet.drawable.sourceRect.h * facet.scale.y);
-			facet.drawable.targetRect.x = position[0] + facet.offset.x;
-			facet.drawable.targetRect.y = position[1] + facet.offset.y - facet.drawable.targetRect.h;
-			facet.drawable.zIndex = position[2];
+			facet->drawable.sourceRect = facet->currentAnimation.frames[facet->currentAnimation.currentFrameIndex].frame;
+			facet->drawable.layerId = system->getDrawableLayerId();
+			facet->drawable.targetRect.w = roundFloat(facet->drawable.sourceRect.w * facet->scale.x);
+			facet->drawable.targetRect.h = roundFloat(facet->drawable.sourceRect.h * facet->scale.y);
+			facet->drawable.targetRect.x = position[0] + facet->offset.x;
+			facet->drawable.targetRect.y = position[1] + facet->offset.y - facet->drawable.targetRect.h;
+			facet->drawable.zIndex = position[2];
 			//facet.drawable.id = single<Renderer>().createDrawable(facet.drawable);
 
-			lua.pushStack(facet.id());
+			lua.pushStack(facet->id());
 		}
 		else {
 			lua.pushStack(-1L);
